tlsf-mode.el: fix highlighting of function parameters
The current anchored-matcher for highlighting the function parameters
is the following:

```
"^\\s-*\\(\\w+\\)\\s-*\\((.*)\\)?\\s-*="
  ("\\<\\w*\\>"
   (progn
     (re-search-backward "(")
     (save-excursion (search-forward-regexp "[^)]*)")))
   nil
   (0 font-lock-constant-face))
```

This does not work, because if the file contains

```
GLOBAL {
  PARAMETERS {
    N = 3;
    M = 3;
  }
  DEFINITIONS {
    Pos(grid, i, j) = grid[i + j * N];
    ...
  }
}
```

The first regex with match `N =` and then `(re-search-backward "(")`
will fail.  This causes font-lock to stop processing the file further.
In other words, the presence of those global parameters in my file
prevents the highlighting of all the arguments of function
definitions.

It gets worse if I want to display some TLSF in some org-mode
document.  Because font-lock fails at some point, org-mode refuses to
highlight TLSF blocks.

This patch improves the existing code by (1) not using
re-search-backward/search-forward-regexp for positions that have
already been matched, (2) dealing with the case where there is no
parenthetical block (i.e., the code matched a parameter).